Move MockMessage from timezone.py into utils
